# base: reduce MakeFilename allocations